--- a/tests/cmdline/params/options/test_callable.py
+++ b/tests/cmdline/params/options/test_callable.py
@@ -7,9 +7,13 @@
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
 """Tests for the :mod:`aiida.cmdline.params.options.callable` module."""
+import sys
+
 import pytest
 from aiida.cmdline.commands.cmd_verdi import verdi
 from click.shell_completion import ShellComplete
+
+SLOW_IMPORTS = ['pydantic']
 
 
 def _get_completions(cli, args, incomplete):
@@ -28,6 +32,14 @@ def unload_config():
     configuration.CONFIG = config
 
 
+@pytest.fixture
+def unimport_slow_imports():
+    """Pretend that pydantic has not been imported yet."""
+
+    for modulename in SLOW_IMPORTS:
+        del sys.modules['pydantic']
+
+
 @pytest.mark.usefixtures('unload_config')
 def test_callable_default_resilient_parsing():
     """Test that tab-completion of ``verdi`` does not evaluate defaults that load the config, which is expensive."""
@@ -39,17 +51,21 @@ def test_callable_default_resilient_parsing():
     assert configuration.CONFIG is None
 
 
-# TODO: I think we'll need to monkeypatch `sys.modules` (is that even possible?)
 @pytest.mark.usefixtures('unload_config')
-def test_undesired_imports_during_tab_completion():
+@pytest.mark.usefixtures('unimport_slow_imports')
+def test_slow_imports_during_tab_completion():
     """Check that verdi does not import certain python modules
     that would make tab-completion slow.
 
     NOTE: This is analogous to `verdi devel check-undesired-imports`
     """
-    import sys
 
-    for modulename in [
-        'pydantic',
-    ]:
-        assert modulename not in sys.modules, f'Detected loaded module {modulename} during tab-completion'
+    # Let's double check that the undesired imports are not already loaded
+    for modulename in SLOW_IMPORTS:
+        assert modulename not in sys.modules, f'Detected loaded module {modulename}'
+
+    completions = [c.value for c in _get_completions(verdi, [], '')]
+    assert 'help' in completions
+
+    for modulename in SLOW_IMPORTS:
+        assert modulename not in sys.modules, f'Detected loaded module {modulename} during tab completion'