--- a/tests/conftest.py
+++ b/tests/conftest.py
@@ -176,8 +176,8 @@ def isolated_config(monkeypatch):
     changing it in tests maybe necessary if a command is invoked that will be reading the config from disk in another
     Python process and so doesn't have access to the loaded config in memory in the process that is running the test.
     """
-    from aiida.manage.configuration.config import Config
     from aiida.manage import configuration
+    from aiida.manage.configuration.config import Config
 
     monkeypatch.setattr(Config, '_backup', lambda *args, **kwargs: None)