--- a/tests/manage/configuration/test_config.py
+++ b/tests/manage/configuration/test_config.py
@@ -14,7 +14,8 @@ import uuid
 
 import pytest
 from aiida.common import exceptions
-from aiida.manage.configuration import Config, Profile, settings
+from aiida.manage.configuration import Profile, settings
+from aiida.manage.configuration.config import Config
 from aiida.manage.configuration.migrations import CURRENT_CONFIG_VERSION, OLDEST_COMPATIBLE_CONFIG_VERSION
 from aiida.manage.configuration.options import get_option
 from aiida.orm.implementation.storage_backend import StorageBackend