`pydantic` adds a lot of time to aiida startup, which is especially detrimental to tab completion.
Here I test the approach of deferring import of `aiida.manage.configuration.config` until really needed, which in turn defers pydantic.

Upcoming usage of pydantic from #6255 seems to be centered in `aiida.orm` which is not imported by default in `verdi` so this PR still seems worth it.

This is in principle a breaking change since we no longer export `Config` in `aiida.manage.configuration` module. But I guess it is not meant to be manipulated directly by users anyway.