@danielhollas 2024-02-04
Some benchmarks using `verdi --version` as a proxy for tab-completion, with pydantic v2.6.0.
I am seeing ~35ms speedup on my new fancy NVMe drive. On my old laptop the difference would likely be much bigger. 

**main**
```python
$ hyperfine -w 5 'verdi --version' 
Benchmark 1: verdi --version
  Time (mean ± σ):     101.4 ms ±   2.3 ms    [User: 83.6 ms, System: 16.5 ms]
  Range (min … max):    97.6 ms … 105.6 ms    30 runs
```

**this PR**
```python
Benchmark 1: verdi --version
  Time (mean ± σ):      65.2 ms ±   2.1 ms    [User: 50.9 ms, System: 13.2 ms]
  Range (min … max):    62.2 ms …  69.8 ms    45 runs
```