@edan-bainglass 2024-02-06
Thank you @danielhollas for addressing this. I have a question - how much does `pydantic` load-time affect AiiDAlab? Is this one of the sources of lag in app load time?