@sphuber 2024-02-07
> CC @sphuber this is ready for review. Happy to hear your thoughts. This PR kind of assumes that the use of pydantic will stay somewhat localized to `aiida.orm` module. If you have other plans than perhaps this is not worth it.

Well, it already _is_ used outside of `aiida.orm`, e.g.:
https://github.com/aiidateam/aiida-core/blob/b7e59a0dbc0dd629be5c8178e98c70e7a2c116e9/src/aiida/storage/psql_dos/backend.py#L16

And it is in `aiida.cmdline`
https://github.com/aiidateam/aiida-core/blob/b7e59a0dbc0dd629be5c8178e98c70e7a2c116e9/src/aiida/cmdline/groups/dynamic.py#L130
but I made sure to keep imports inside methods whenever possible. But I am not a 100% sure if this is not still evaluated during tab-completion, because one of the main motivations for adopting `pydantic` is to have `verdi` add subcommands dynamically based on which plugins are installed. See https://github.com/aiidateam/aiida-core/pull/6190

In https://github.com/aiidateam/aiida-core/pull/6255 we go even further and essentially use `pydantic` in almost each `aiida.orm` submodule. Would those automatically be handled by the changes in this PR? Or would special care still have to be taken? Is there anyway we can test this for regression reliably. There currently is a very basic test in the CI that we added years back, which simply measures the run time of `verdi` in a loop. But it is quite fragile if we set the limit too close to the ideal time (which is also strongly system specific) and so we risk false positives.

> This is in principle a breaking change since we no longer export Config in aiida.manage.configuration module. But I guess it is not meant to be manipulated directly by users anyway.

Maybe not users, but third-party applications may very well be using it directly. For example, would AiiDAlab maybe not be affected? Surely they will operate on the `Config`, for example to set options etc and get profile information. Is this the only way of deferring the import of pydantic?