@danielhollas 2024-02-07
Thanks for taking a look @sphuber!

> Well, it already is used outside of aiida.orm, e.g.:

Right, but I indeed double checked that in those cases the imports are inlined so I assumed it will be fine.
I have now verified this assumption manually, that `pydantic` is indeed not being imported during tab-completion (by adding `raise ValueError` in `pydantic/__init__.py`. 

> Would those automatically be handled by the changes in this PR?

We already verify that `aiida.orm` is not being imported in verdi startup, by running `verdi devel check-load-time` so provided that you didn't add pydantic import outside of `aiida.orm` that PR should be fine. 


> Is there anyway we can test this for regression reliably.

Indeed we can, that's what `verdi devel check-undesired-imports` is for! Unfortunately, it does not work in this case, because actually running `verdi devel` loads the configuration. In the tab-completion case, we have a special monkey-patch that prevents evaluation of dynamic default values that I added in https://github.com/aiidateam/aiida-core/pull/6144.

So instead I've added a regression test that calls tab-completion internally.

> Maybe not users, but third-party applications may very well be using it directly. For example, would AiiDAlab maybe not be affected?

I don't think so? In `AiiDAlab` we mostly only use `load_profile`. I would assume third-party code would similarly only use the helper functions defined in `aiida/manage/configuration/__init__.py` such as `get_config() / load_config()`. 

> Is this the only way of deferring the import of pydantic?

Not sure how to answer this. :sweat_smile: The `Config` class is derived from the pydantic BaseClass, and so `pydantic` needs to be imported when the `Config` class itself (not its instances) is built. I don't see a way how to make the class Config available in `aiida.manage` while also not importing `pydantic`.