@danielhollas 2024-02-08
By the way, I've just found out about `verdi devel play`, what a beautiful easter egg. :joy: :notes: