@sphuber 2024-02-08
> By the way, I've just found out about `verdi devel play`, what a beautiful easter egg. 😂 🎶

It's not the only easter egg in `verdi` (hint, hint) ;)