@danielhollas 2024-02-08
> Very nice. I am wondering if the verdi devel check-undesired-imports is now superfluous as it is a subset of the unit test you added? I think technically there is still a difference in code path between merely tab-completing, and actually invoking a command. As you say, some of the command parameters were changed to lazily evaluate their default value, such that they are only executed when the command is actually called. What we are really interested in is that the tab-complete is responsive. So I think we can just add the modules from verdi devel check-undesired-imports to your unit test and get rid of that command (and invocation in the CI scripts). Right?

I've been thinking about this as well. As you mention, `verdi devel check-undesired-imports` provide stronger guarantees. I've looked at the list of the blacklisted modules and for most of them I think this stronger guarantee is warranted, since many of them incur significant import cost. In other words, I think it's reasonable to expect that verdi commands that can be fast are actually fast (e.g. `asyncio` should be loaded only when really needed).